Download error hyperlinks and cleaning up CoreDownloadItem

+++ components/core-api/src/main/java/org/limewire/core/api/download/DownloadItem.java
@@ -32,12 +32,12 @@ public interface DownloadItem  {
 	/**
 	 * @return size in bytes
 	 */
-	public double getCurrentSize();
+	public long getCurrentSize();
 	
 	/**
      * @return size in bytes
      */
-	public double getTotalSize();
+	public long getTotalSize();
 
 	public String getRemainingDownloadTime();
